# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ListFlow AI Resale Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List all marketplace listings
      parameters:
        - name: itemId
          in: query
          schema:
            type: integer
        - name: marketplace
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, ended, sold, draft]
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a listing for an item on a marketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a single listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingPatch"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, awaiting_shipment, shipped, delivered, returned, refunded]
        - name: marketplace
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (record a sale)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPatch"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  # ── Shipping ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Listing:
      type: object
      required: [id, itemId, marketplace, status, price, createdAt, updatedAt]
      properties:
        id:
          type: integer
        itemId:
          type: integer
        marketplace:
          type: string
          enum: [ebay, poshmark, depop, mercari, facebook, etsy, grailed, whatnot, shopify]
        status:
          type: string
          enum: [draft, active, ended, sold]
        price:
          type: number
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        marketplaceListingId:
          type: ["string", "null"]
        listedAt:
          type: ["string", "null"]
          format: date-time
        soldAt:
          type: ["string", "null"]
          format: date-time
        itemTitle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListingInput:
      type: object
      required: [itemId, marketplace, price, status]
      properties:
        itemId:
          type: integer
        marketplace:
          type: string
          enum: [ebay, poshmark, depop, mercari, facebook, etsy, grailed, whatnot, shopify]
        status:
          type: string
          enum: [draft, active, ended, sold]
        price:
          type: number
          minimum: 0
        title:
          type: string
        description:
          type: string
        marketplaceListingId:
          type: string
    ListingPatch:
      type: object
      properties:
        marketplace:
          type: string
          enum: [ebay, poshmark, depop, mercari, facebook, etsy, grailed, whatnot, shopify]
        status:
          type: string
          enum: [draft, active, ended, sold]
        price:
          type: number
          minimum: 0
        title:
          type: string
        description:
          type: string
        marketplaceListingId:
          type: string
        listedAt:
          type: string
          format: date-time
        soldAt:
          type: string
          format: date-time
...
    Order:
      type: object
      required: [id, itemId, marketplace, salePrice, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        itemId:
          type: integer
        listingId:
          type: ["integer", "null"]
        marketplace:
          type: string
          enum: [ebay, poshmark, depop, mercari, facebook, etsy, grailed, whatnot, shopify]
        buyerName:
          type: ["string", "null"]
        salePrice:
          type: number
        fees:
          type: number
        shippingCost:
          type: number
        profit:
          type: number
        status:
          type: string
          enum: [pending, awaiting_shipment, shipped, delivered, returned, refunded]
        trackingNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        itemTitle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [itemId, marketplace, salePrice, status]
      properties:
        itemId:
          type: integer
        listingId:
          type: integer
        marketplace:
          type: string
          enum: [ebay, poshmark, depop, mercari, facebook, etsy, grailed, whatnot, shopify]
        buyerName:
          type: string
        salePrice:
          type: number
          minimum: 0
        fees:
          type: number
          minimum: 0
        shippingCost:
          type: number
          minimum: 0
        profit:
          type: number
        status:
          type: string
          enum: [pending, awaiting_shipment, shipped, delivered, returned, refunded]
        trackingNumber:
          type: string
        notes:
          type: string
    OrderPatch:
      type: object
      properties:
        buyerName:
          type: string
        salePrice:
          type: number
          minimum: 0
        fees:
          type: number
          minimum: 0
        shippingCost:
          type: number
          minimum: 0
        profit:
          type: number
        status:
          type: string
          enum: [pending, awaiting_shipment, shipped, delivered, returned, refunded]
        trackingNumber:
          type: string
        notes:
          type: string

    # ── Shipping ──────────────────────────────────────────────────────────────
...