# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ListFlow AI Resale Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingInputMarketplace } from './listingInputMarketplace';
import type { ListingInputStatus } from './listingInputStatus';

export interface ListingInput {
  itemId: number;
  marketplace: ListingInputMarketplace;
  status: ListingInputStatus;
  /** @minimum 0 */
  price: number;
  title?: string;
  description?: string;
  marketplaceListingId?: string;
}