# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ListFlow AI Resale Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputMarketplace } from './orderInputMarketplace';
import type { OrderInputStatus } from './orderInputStatus';

export interface OrderInput {
  itemId: number;
  listingId?: number;
  marketplace: OrderInputMarketplace;
  buyerName?: string;
  /** @minimum 0 */
  salePrice: number;
  /** @minimum 0 */
  fees?: number;
  /** @minimum 0 */
  shippingCost?: number;
  profit?: number;
  status: OrderInputStatus;
  trackingNumber?: string;
  notes?: string;
}