# lib/api-zod/src/generated/types/orderPatch.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ListFlow AI Resale Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPatchStatus } from './orderPatchStatus';

export interface OrderPatch {
  buyerName?: string;
  /** @minimum 0 */
  salePrice?: number;
  /** @minimum 0 */
  fees?: number;
  /** @minimum 0 */
  shippingCost?: number;
  profit?: number;
  status?: OrderPatchStatus;
  trackingNumber?: string;
  notes?: string;
}